Add stub for testing the Valohai API
Testing calling the API from inside an execution.
The first step is to get the API URL from config.
It is not directly available, so we parse it from an URL that does
exist.

To be able to include the helper, the step command adds the current path
to PYTHONPATH -- it is not pretty, but this way no TOML configuration is
required (as this is not a coherent project but instead a bunch of
separate scripts to run as steps).