Keep directories in step 'log-inputs'
Add the 'keep-directories: full' property to the step's 'input_data'
input. This is to test giving the step multiple inputs with the same
filename but different paths.

By default, these are not separated, so the script will only get one
actual input, but with keep-inputs, the path is preserved so the file
paths will be different, and both files are handled.

This will be used to test how this handling works when the input is a
dataset version that contains more than one file with the same name but
different path.